Added a nickname changer under the settings menu
Allows the user to chnage their name, limits them to not allow a blank name or if they try to enter their same name.
Known issue - colours for errors are different this is due to one being send client side and one being sent from  the server. There is currently no support for sending colour in packets